empire/x-analyzer/requirements.txt
redis[hiredis]==5.2.1
anthropic==0.42.0
yt-dlp==2024.12.23
openai-whisper==20240930
aiohttp==3.11.11
beautifulsoup4==4.12.3
selenium==4.27.1
python-dotenv==1.0.1
